requirements.txt
# Core dependencies
torch>=1.12.0
dgl>=0.9.0
numpy>=1.21.0
pandas>=1.3.0

# Machine learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.8.0

# Bioinformatics
biopython>=1.79

# Optional: ESM2 support
# fair-esm>=2.0.0

# Optional: Additional tools
tqdm>=4.62.0

